+++ src/assets/metadata/index.js
@@ -1,0 +1,7 @@
+import commonAttributes from 'src/assets/metadata/common_attributes.json';
+import apiVersions from 'src/assets/metadata/apiVersions';
+
+export default {
+  commonAttributes,
+  apiVersions,
+};