+++ changelog.md
@@ -9,8 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Setup project
-- First version of Kubernetes parser
+- Setup project.
+- Add metadata for main Kubernetes resources.
+- Add parser for Kubernetes files.
+- Add renderer for Kubernetes files.
 
 ### Changed